refactor: isZoomBarEnabled(), isToolbarEnabled() in zoom service
